app/requirements.txt
Flask~=2.3.2
gunicorn~=23.0.0
requests~=2.31.0
pyyaml~=6.0.2
pydantic~=2.11.5